+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 ## Changelog
 
-Here goes the changelog text.
+0.1.2
+
+- Feature: Add command line interface to export data
+```bash
+$ bin/console db:dump
+$ bin/console db:import
+$ bin/console db:drop
+$ bin/console db:create
+```
+- Feature: Add support for exporting data as a tar.gz file
+- Feature: Add support for cleaning up export folder
+